+++ oap-server/server-receiver-plugin/envoy-metrics-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/envoy/als/AbstractALSAnalyzer.java
@@ -33,9 +33,6 @@ public abstract class AbstractALSAnalyzer implements ALSHTTPAnalysis {
             return defaultRole;
         }
         final Node node = alsIdentifier.getNode();
-        if (node == null) {
-            return defaultRole;
-        }
         final String id = node.getId();
         if (id.startsWith("router~")) {
             return Role.PROXY;